requirements.txt
steam>=1.4.4
requests>=2.31.0
python-dotenv>=1.0.0